cont: collapsing device management routes into /admin/devices
