+++ ui/user/src/lib/components/admin/McpServerEntryForm.svelte
@@ -157,17 +157,6 @@
 	});
 	let configurationReadonly = $derived(readonly || isCatalogEntryDeployedMultiUserServer(entry));
 
-	type ScrollState = {
-		hasMoreLeft: boolean;
-		hasMoreRight: boolean;
-		ref?: HTMLDivElement;
-	};
-	let rootScrollContainer = $state<ScrollState>({
-		hasMoreLeft: false,
-		hasMoreRight: false,
-		ref: undefined
-	});
-
 	let oauthDialog = $state<ReturnType<typeof ResponsiveDialog>>();
 	let oauthURL = $state<string>();
 	let oauthURLs = $state<Record<string, string>>();
@@ -391,13 +380,7 @@
 			users = data;
 		});
 
-		checkScrollPosition();
-		rootScrollContainer.ref?.addEventListener('scroll', checkScrollPosition);
-		window.addEventListener('resize', checkScrollPosition);
-
 		return () => {
-			rootScrollContainer.ref?.removeEventListener('scroll', checkScrollPosition);
-			window.removeEventListener('resize', checkScrollPosition);
 			document.removeEventListener('visibilitychange', handleVisibilityChange);
 		};
 	});
@@ -437,26 +420,6 @@
 			item.serverUserType === 'multiUser' &&
 			!!item.catalogEntryID
 		);
-	}
-
-	function checkScrollPosition() {
-		if (!rootScrollContainer.ref) return;
-
-		const { scrollLeft, scrollWidth, clientWidth } = rootScrollContainer.ref;
-		rootScrollContainer.hasMoreLeft = scrollLeft > 0;
-		rootScrollContainer.hasMoreRight = scrollLeft < scrollWidth - clientWidth - 1; // -1 for rounding errors
-	}
-
-	function scrollLeft() {
-		if (rootScrollContainer.ref) {
-			rootScrollContainer.ref.scrollBy({ left: -200, behavior: 'smooth' });
-		}
-	}
-
-	function scrollRight() {
-		if (rootScrollContainer.ref) {
-			rootScrollContainer.ref.scrollBy({ left: 200, behavior: 'smooth' });
-		}
 	}
 
 	function handleSelectionChange(newSelection: string) {
@@ -857,13 +820,12 @@
 			<OverflowContainer
 				class="scrollbar-none flex min-h-12 w-full items-center gap-2 overflow-x-auto"
 				style="scroll-behavior: smooth;"
-				{@attach (node: HTMLDivElement) => (rootScrollContainer.ref = node)}
 			>
-				{#snippet children({ x })}
+				{#snippet children({ x, hasMoreLeft, hasMoreRight, scrollLeft, scrollRight })}
 					{#if tabs.length > 0 && (entry?.id || server?.id)}
 						{#if x}
 							<button
-								disabled={!rootScrollContainer.hasMoreLeft}
+								disabled={!hasMoreLeft}
 								onclick={scrollLeft}
 								class="bg-base-200 dark:bg-base-100 sticky left-0 flex aspect-square h-full items-center justify-center rounded-l-md p-2.5 opacity-100 transition-all duration-200 disabled:opacity-30"
 							>
@@ -904,7 +866,7 @@
 
 						{#if x}
 							<button
-								disabled={!rootScrollContainer.hasMoreRight}
+								disabled={!hasMoreRight}
 								onclick={scrollRight}
 								class="bg-base-200 dark:bg-base-100 sticky right-0 flex aspect-square h-full items-center justify-center rounded-r-md p-2.5 opacity-100 transition-all duration-200 disabled:opacity-30"
 							>
