+++ ui/user/src/lib/components/admin/SetupSplashDialog.svelte
@@ -13,6 +13,10 @@
 
 	let dialog = $state<ReturnType<typeof ResponsiveDialog>>();
 	let loading = $state(false);
+
+	const authProviderPath = '/admin/auth-providers';
+	const modelProviderPath = '/admin/model-providers';
+
 	const storeData = $derived($adminConfigStore);
 	const isAuthProviderConfigured = $derived(
 		version.current.authEnabled ? storeData.authProviderConfigured : true
@@ -20,7 +24,7 @@
 	const requiresModelProviderConfiguration = $derived(
 		version.current.agentsEnabled !== false && !storeData.modelProviderConfigured
 	);
-	const isOnAuthProvidersPage = $derived(page.url.pathname === '/admin/auth-providers');
+	const isOnAuthProvidersPage = $derived(page.url.pathname === authProviderPath);
 	const isBootstrapUser = $derived(profile.current.isBootstrapUser?.() ?? false);
 
 	$effect(() => {
@@ -110,9 +114,9 @@
 				}
 
 				if (!isAuthProviderConfigured) {
-					goto('/admin/auth-providers');
-				} else {
-					goto('/admin/model-providers');
+					goto(authProviderPath);
+				} else if (requiresModelProviderConfiguration) {
+					goto(modelProviderPath);
 				}
 				dialog?.close();
 			}}
@@ -129,11 +133,10 @@
 			onclick={() => {
 				handleAcceptEula();
 				localStorage.setItem('seenSplashDialog', new Date().toISOString());
-				if (storeData.modelProviderConfigured) {
-					dialog?.close();
-				} else {
-					goto('/admin/model-providers');
+				if (requiresModelProviderConfiguration && page.url.pathname !== modelProviderPath) {
+					goto(modelProviderPath);
 				}
+				dialog?.close();
 			}}
 		>
 			Continue
