+++ ui/user/src/routes/admin/device-clients/+page.ts
@@ -2,5 +2,5 @@ import type { PageLoad } from './$types';
 import { redirect } from '@sveltejs/kit';
 
 export const load: PageLoad = async () => {
-	throw redirect(301, '/admin/device-overview?view=device-clients');
+	throw redirect(301, '/admin/devices?view=device-clients');
 };
