+++ ui/user/src/routes/admin/device-skills/[name]/+page.svelte
@@ -79,7 +79,7 @@
 		if (typeof window !== 'undefined' && window.history.length > 1) {
 			window.history.back();
 		} else {
-			goto(resolve('/admin/device-skills'));
+			goto(resolve('/admin/devices?view=device-skills'));
 		}
 	}}
 >
